requirements.txt
aiogram==2.24
asyncio
requests
datetime
python-decouple
pandas
# dotenv
# os